requirements.txt
Flask>=2.0
PyJWT>=2.0
flask-pymongo>=2.3
pymongo>=4.0
python-dotenv>=1.0
Flask-Cors>=6.0.0
gunicorn>=20.1
